Upgraded webpack
