+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "start": "./bin/cli serve --jobs",
     "dev-start": "nodemon --config .nodemon.json --exec \"./bin/cli -c .env serve --jobs\"",
-    "build": "NODE_ENV=production webpack --config webpack.config.js --bail",
-    "build-watch": "NODE_ENV=development webpack --config webpack.config.dev.js --watch",
+    "build": "NODE_ENV=production webpack -p --config webpack.config.js --bail",
+    "build-watch": "NODE_ENV=development webpack --config webpack.config.js --watch",
     "lint": "eslint bin/* .",
     "lint-fix": "eslint bin/* . --fix",
     "test": "TEST_MODE=unit NODE_ENV=test mocha -R ${NPM_PACKAGE_CONFIG_MOCHA_REPORTER:-spec}",
@@ -159,7 +159,7 @@
     "style-loader": "^0.13.1",
     "supertest": "^2.0.1",
     "timeago.js": "^2.0.3",
-    "webpack": "^1.13.3"
+    "webpack": "^2.2.1"
   },
   "engines": {
     "node": ">=7.0.0"
